[ci skip] 4677039137?check_suite_focus=true https://github.com/tkoyama010/pyvista-docs-dev-ja/runs/4677039137?check_suite_focus=true

+++ api/core/_autosummary/pyvista-PolyData-clean-1.py
@@ -3,7 +3,7 @@
 #
 import pyvista as pv
 import numpy as np
-points = np.array([[0, 0, 0], [0, 1, 0], [1, 0, 0]])
+points = np.array([[0, 0, 0], [0, 1, 0], [1, 0, 0]], dtype=np.float32)
 faces = np.array([3, 0, 1, 2, 3, 0, 3, 3])
 mesh = pv.PolyData(points, faces)
 mout = mesh.clean()
