+++ api/core/_autosummary/pyvista-StructuredGrid-1.py
@@ -13,8 +13,8 @@ grid = pyvista.StructuredGrid(vtkgrid)
 #
 # Create from NumPy arrays
 #
-xrng = np.arange(-10, 10, 2)
-yrng = np.arange(-10, 10, 2)
-zrng = np.arange(-10, 10, 2)
+xrng = np.arange(-10, 10, 2, dtype=np.float32)
+yrng = np.arange(-10, 10, 2, dtype=np.float32)
+zrng = np.arange(-10, 10, 2, dtype=np.float32)
 x, y, z = np.meshgrid(xrng, yrng, zrng)
 grid = pyvista.StructuredGrid(x, y, z)
