+++ api/core/_autosummary/pyvista-StructuredGrid-x-1.py
@@ -1,8 +1,8 @@
 import pyvista
 import numpy as np
-xrng = np.arange(-10, 10, 1)
-yrng = np.arange(-10, 10, 2)
-zrng = np.arange(-10, 10, 5)
+xrng = np.arange(-10, 10, 1, dtype=np.float32)
+yrng = np.arange(-10, 10, 2, dtype=np.float32)
+zrng = np.arange(-10, 10, 5, dtype=np.float32)
 x, y, z = np.meshgrid(xrng, yrng, zrng)
 grid = pyvista.StructuredGrid(x, y, z)
 grid.x.shape
